Crack ContentArea view into pieces (specific components)
